+++ R/ggh4x-package.R
@@ -15,6 +15,7 @@ NULL
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
+#' @importFrom cli cli_abort
 #' @importFrom lifecycle deprecated
 #' @importFrom stats median
 ## usethis namespace: end